+++ backend/schemas/advance_schema.py
@@ -1,4 +1,5 @@
 from pydantic import BaseModel
+from decimal import Decimal
 
 class AdvanceRequest(BaseModel):
     user_id: int
@@ -6,8 +7,8 @@ class AdvanceRequest(BaseModel):
     country: str
     currency_from: str
     currency_to: str
-    amount_from: float
-    amount_to: float
-    amount: float
-    market_rate: float
+    amount_from: Decimal
+    amount_to: Decimal
+    amount: Decimal
+    market_rate: Decimal
     pin: str
